Rename: program_path.{h,cpp} => program_dirs.{h,cpp}
