+++ src/HealthGPS/CMakeLists.txt
@@ -77,8 +77,8 @@ target_sources(
             "person.h"
             "population.cpp"
             "population.h"
-            "program_path.cpp"
-            "program_path.h"
+            "program_dirs.cpp"
+            "program_dirs.h"
             "physical_activity_scenario.cpp"
             "physical_activity_scenario.h"
             "random_algorithm.cpp"
